perf: faster tests (#8113)
* refactor: make libcurl thread shutdown faster during tests

* docs: fix out-of-date code comment in web.h

* refactor: tr_sessionClose() now takes fractions of seconds for deadline

refactor: make tr_web::startShutdown() deadline dependent on the tr_sessionClose() deadline arg

* test: set tr_sessionClose() deadline of 0.5 secs

* refactor: FileTest now inherits from SandboxedTest

* fixup! test: set tr_sessionClose() deadline of 0.5 secs

+++ libtransmission/transmission.h
@@ -232,7 +232,7 @@ void tr_sessionReloadBlocklists(tr_session* session);
  *
  * @param timeout_secs specifies how long to wait on these announces.
  */
-void tr_sessionClose(tr_session* session, size_t timeout_secs = 15);
+void tr_sessionClose(tr_session* session, double timeout_secs = 15.0);
 
 /**
  * @brief Return the session's configuration directory.
